Create README.md
The three R scripts are slightly different version of the simulation. The first script calculates the new values for ech locality by multiplying the presnt number of individuals by the fitness value. This behaviour can cause the effect that at certain localities the populations do not reach the full capacity, if the fitness value for both species are to low. The second version accelerates the growth, even if the local fitness is low, in the case that the present number of individuals is below the full capacity by introducing an correction factor. The third script multiplies the r-values by the fitness value. This way a low fitness value does not reduce the number of individuals unless the compiting species has an advantage caused by higher spore arrival or higher fitness value.